--- src/components/CreatePost/setDate/setDate.style.ts
@@ -1,20 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  height: 90%;
-
-  display: flex;
-  flex-direction: column;
-  justify-content: space-around;
-`;
-
-export const DateStringContainer = styled.div`
-  padding: 10px;
-
-  text-align: right;
-  font-weight: var(--weight-semi-bold);
-`;
-
-export const DatePickerContainer = styled.div`
-  margin-bottom: 300px;
-`;
