+++ src/domains/CreatePost/Page/SetPlaceMapPage/SetPlaceMap.style.ts
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const SubmitContainer = styled.div`
+  padding: 30px 20px;
+`;
