+++ src/domains/CreatePost/Page/SetPlacePage/SetPlace.style.ts
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+export const MyLocationButton = styled.button`
+  margin-top: 30px;
+  margin-left: auto;
+  display: flex;
+  align-items: center;
+
+  color: var(--color-blue);
+  font-weight: var(--weight-semi-bold);
+
+  svg {
+    width: 20px;
+    height: 20px;
+    margin-right: 5px;
+  }
+`;
