+++ src/domains/CreatePost/components/SubmitButton.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const SubmitButton = styled.button`
+  width: 100%;
+  padding: 12px;
+
+  border-radius: 6px;
+  background-color: var(--color-main);
+
+  color: var(--color-white);
+  font-weight: var(--weight-bold);
+  text-align: center;
+`;
