+++ src/domains/CreatePost/constants/Place.ts
@@ -1,0 +1,10 @@
+export const PLACE_KEYWORDS = {
+  origin: {
+    keyWord: '출발지',
+    content: '여기에서 출발',
+  },
+  destination: {
+    keyWord: '도착지',
+    content: '여기로 도착',
+  },
+};
