+++ src/domains/CreatePost/utils/setDepartureTimeValueFunc.ts
